disas/riscv: Reduce rv_insn to uint32_t
Since 758dce9c98a, the only possible values for
instruction length are 2 and 4.

Message-ID: <20260812223142.349142-3-richard.henderson@linaro.org>

+++ disas/riscv.h
@@ -11,7 +11,7 @@
 
 /* types */
 
-typedef uint64_t rv_inst;
+typedef uint32_t rv_inst;
 typedef uint16_t rv_opcode;
 
 /* enums */
